requirements.txt
absl-py==1.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
imageio==2.22.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
networkx==2.8.7
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
split-folders==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.8.12
torch==1.12.1+cu116
torch-tb-profiler==0.4.0
torchaudio==0.12.1+cu116
torchmetrics==0.10.0
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
